+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Change log
 - Changed ``CabinetForeignKey`` to reference the configured file model
   by default.
 - Limited the maximum width of the inline folder select widget.
+- Added tests for the ``CabinetForeignKey``.
 
 
 `0.8`_ (2018-12-14)